Use celery in CategoryResults
